Pass version to install command as to not use find
Since now we have the full name of the version we don't need to use find
to find (hehe) the downloaded archive, we just use the version prepend
with `kafka_`, e.g. `kafka_$version`.

This prevent a whole class of bugs, specially when we are talking about differences
between the commands in GNU/Linux system vs BSD based ones (MacOS).

One of those bugs, is that BSD find does not support the `-name` flag,
which was erroring out. Because of that, the `$dist` variable was not
rightfully populated, which could be troublesome since in the next line
we were doing a `mv $dist/* .` and a `rf -r $dist`, both of which are
destructive actions.